Add optional legend title (#105)
This pull request introduces enhancements to the `Legend` functionality
in the `egui_plot` library, updates the plot demos to include legend
titles, and modifies related tests and snapshots to reflect these
changes. The most significant changes include adding a `title` property
to the `Legend` struct, updating plot demos to utilize this new feature,
and adjusting test infrastructure and snapshots accordingly.

### Enhancements to `Legend` functionality:
*
[`egui_plot/src/legend.rs`](diffhunk://#diff-41de195c6ebce678a328dcb6d36642dcccd13eb73aa023e13590b61e2f74f1cfR49):
Added a `title` property to the `Legend` struct, with a default value of
`None`. Introduced a new `title` method to set the legend title and
updated the `LegendWidget` to display the title if it is set.
[[1]](diffhunk://#diff-41de195c6ebce678a328dcb6d36642dcccd13eb73aa023e13590b61e2f74f1cfR49)
[[2]](diffhunk://#diff-41de195c6ebce678a328dcb6d36642dcccd13eb73aa023e13590b61e2f74f1cfR64)
[[3]](diffhunk://#diff-41de195c6ebce678a328dcb6d36642dcccd13eb73aa023e13590b61e2f74f1cfR94-R100)
[[4]](diffhunk://#diff-41de195c6ebce678a328dcb6d36642dcccd13eb73aa023e13590b61e2f74f1cfR340-R342)

### Updates to plot demos:
*
[`demo/src/plot_demo.rs`](diffhunk://#diff-5a1025d5dd6c016d4b4017e33bb3322bd767ab0591195efa41a2a41da09efc9eL286-R286):
Updated the `Legend` configuration in `LineDemo`, `MarkerDemo`, and
`ItemsDemo` to include titles ("Lines", "Markers", and "Items",
respectively).
[[1]](diffhunk://#diff-5a1025d5dd6c016d4b4017e33bb3322bd767ab0591195efa41a2a41da09efc9eL286-R286)
[[2]](diffhunk://#diff-5a1025d5dd6c016d4b4017e33bb3322bd767ab0591195efa41a2a41da09efc9eL376-R376)
[[3]](diffhunk://#diff-5a1025d5dd6c016d4b4017e33bb3322bd767ab0591195efa41a2a41da09efc9eL764-R768)

* Closes <https://github.com/emilk/egui_plot/issues/98>
* [x] I have followed the instructions in the PR template

---------

Co-authored-by: lucasmerlin <hi@lucasmerlin.me>