Fix BoxPlot legend (#97)
Fvg on discord noticed that BoxPlots were missing the legend. 

![Screenshot from 2025-04-07
23-16-30](https://github.com/user-attachments/assets/9ab9fbb6-5fb3-4683-899d-91b323e4bf60)


Upon further investigation, either some merge was a bit icky, or
something was missed when adding in `PlotItemBase`.

To fix this i have

* **Replaced** `BoxPlot::color` with `BoxPlot::default_color`, as it was
a name collision with `PlotItem::color`.
* **Replaced** `BarChart::color` with `BarChart::default_color`, as it
was a name collision with `PlotItem::color`.

I know this is a breaking change without a deprication first, but i feel
this is better than ~~going for a deprication first~~ keeping that name
collision in for longer.

* Remove BoxPlots implementation of the `PlotItem::name()` function, and
the `BoxPlot::name` field, as it was incomplete, and is also covered in
`PlotItemBase`
* Reduced the visibility of `BarChart::default_color` and
`BoxPlot::default_color`, as it is only read, and only in `plot_ui.rs`,
and that can be done via our newly not-name-colliding `PlotItem::color`.

I also noticed there are missing snapshots for the (Charts.png is only
the histogram, and only vertical). Barchart and boxplot are missing. But
that is worth its own PR.

In any case, with this, i get the expected legend back: 

![Screenshot from 2025-04-07
23-16-11](https://github.com/user-attachments/assets/24bde09d-5f59-4dff-be54-aad1de827fb0)


* [x] I have followed the instructions in the PR template
* [x] I have run `check.sh` locally and it went through.

---------

Co-authored-by: Lucas Meurer <lucasmeurer96@gmail.com>
Co-authored-by: lucasmerlin <hi@lucasmerlin.me>