Clarify the relationship between name and id
The name passed to new() defines the initial id, but changing it with
name() does not change the id, it has to be changed explicitly with
id(). This documentation update attempts to make this caveat more
explicit, as it is not apparent from the public API without looking at
the source.