Fix border drawing for spans (#203)
I believe this should fix the border drawing for spans. It got broken
when I added support for (half-)infinite spans.
This should also prevent spans from being drawn altogether when they are
not in the plot bounds. Not sure if that has a positive performance
impact but it can't hurt :slightly_smiling_face:

* Closes #201

---------

Co-authored-by: Michal Sustr <michal.sustr@gmail.com>